+++ R/multistep_xgboost.R
@@ -538,6 +538,13 @@ predict.xgboost_multistep_fit_impl <- function(object, new_data, ...) {
 #' @export
 xgboost_multistep_predict_impl <- function(object, new_data, ...) {
 
+  # Date Mapping Table
+  date_tbl <- new_data %>%
+    dplyr::select(Date, Date_index.num) %>%
+    dplyr::distinct() %>%
+    dplyr::arrange(Date) %>%
+    dplyr::mutate(Run_Number = dplyr::row_number())
+
   # PREPARE INPUTS
   xreg_recipe <- object$extras$xreg_recipe
   h_horizon <- nrow(new_data)
@@ -546,14 +553,16 @@ xgboost_multistep_predict_impl <- function(object, new_data, ...) {
   xreg_tbl <- modeltime::bake_xreg_recipe(xreg_recipe,
     new_data,
     format = "tbl"
-  )
+  ) %>%
+    dplyr::left_join(date_tbl, by = "Date_index.num") %>%
+    dplyr::mutate(Row_Num = dplyr::row_number())
 
   # PREDICTIONS
-  final_prediction <- c()
+  final_prediction <- tibble::tibble()
   start_val <- 1
 
   for (model_name in names(object$models)) {
-    if (start_val > nrow(xreg_tbl)) {
+    if (start_val > nrow(date_tbl)) {
       break
     }
 
@@ -561,9 +570,14 @@ xgboost_multistep_predict_impl <- function(object, new_data, ...) {
 
     xgboost_model <- object$models[[model_name]]
 
-    xreg_tbl_final <- xreg_tbl %>%
-      dplyr::select(tidyselect::any_of(xgboost_model$feature_names)) %>%
-      dplyr::slice(start_val:lag_number)
+    xreg_tbl_temp <- xreg_tbl %>%
+      dplyr::filter(
+        Run_Number >= start_val,
+        Run_Number <= lag_number
+      )
+
+    xreg_tbl_final <- xreg_tbl_temp %>%
+      dplyr::select(tidyselect::any_of(xgboost_model$feature_names))
 
     if (!is.null(xreg_tbl)) {
       preds_xgboost <- modeltime::xgboost_predict(xgboost_model,
@@ -574,9 +588,16 @@ xgboost_multistep_predict_impl <- function(object, new_data, ...) {
       preds_xgboost <- rep(0, h_horizon)
     }
 
+    preds_xgboost <- tibble::tibble(.pred = preds_xgboost) %>%
+      dplyr::mutate(Row_Num = xreg_tbl_temp$Row_Num)
     start_val <- as.numeric(lag_number) + 1
-    final_prediction <- c(final_prediction, preds_xgboost)
+    final_prediction <- rbind(final_prediction, preds_xgboost)
   }
+
+  # Ensure it's sorted correctly for global models
+  final_prediction <- final_prediction %>%
+    dplyr::arrange(Row_Num) %>%
+    dplyr::pull(.pred)
 
   return(final_prediction)
 }